Move code of sheet toolbar buttons to the sheet toolbar module.
